+++ src/entries.test.ts
@@ -20,6 +20,7 @@ import * as v24 from "./v24";
 import * as v25 from "./v25";
 import * as v26 from "./v26";
 import * as v27 from "./v27";
+import * as v28 from "./v28";
 import * as v9 from "./v9";
 
 /**
@@ -50,6 +51,7 @@ const ENTRIES: Record<number, { BYPP_FORMAT_VERSION: number }> = {
   25: v25,
   26: v26,
   27: v27,
+  28: v28,
 };
 
 const SRC = dirname(fileURLToPath(import.meta.url));